refactor(NLP Analysis): analyze.js holds the analyzeSpeech function which returns response obj // {sentiment: 1.4, entities: [lion, mouse]}
